refector: 공식과 신속 분리 작업 진행
