+++ src/entities/stage/create/constants/sportTypes.ts
