+++ src/entities/stage/create/index.ts
@@ -1,0 +1,4 @@
+export { default as SelectStageType } from './ui/SelectStageType';
+export { default as AddButton } from './ui/AddButton';
+export { default as GameField } from './ui/GameField';
+export { categoryTypes, getCategoryLabel } from './constants/sportTypes';
