--- src/entities/stage/create/official/index.ts
@@ -1,2 +1,0 @@
-export { default as SelectStageType } from './ui/SelectStageType';
-export { default as AddButton } from './ui/AddButton';
