+++ src/entities/stage/create/ui/SelectStageType/index.tsx
