+++ src/views/stage/create/fast/index.ts
@@ -1,0 +1,1 @@
+export { default as FastCreatePage } from './ui/FastCreatePage';
