+++ src/widgets/stage/create/constants/miniGames.ts
