+++ src/widgets/stage/create/constants/storeItems.ts
