+++ src/widgets/stage/create/model/miniGameUtils.ts
@@ -1,0 +1,19 @@
+import { UseFormSetValue } from 'react-hook-form';
+import { StageData } from '@/shared/types/stage/create';
+import { MiniGameType } from '../constants/miniGames';
+
+export const getIconColor = (isActive: boolean) =>
+  isActive ? '#526FFE' : '#898989';
+
+export const toggleMiniGame = (
+  game: MiniGameType,
+  isActive: boolean,
+  setValue: UseFormSetValue<StageData>,
+) => {
+  if (isActive) {
+    setValue(`miniGame.${game}.maxBettingPoint`, null);
+    setValue(`miniGame.${game}.minBettingPoint`, null);
+    setValue(`miniGame.${game}.initialTicketCount`, null);
+  }
+  setValue(`miniGame.${game}.isActive`, !isActive);
+};
