--- src/widgets/stage/create/official/model/validateFields.ts
@@ -1,40 +1,0 @@
-import { UseFormWatch } from 'react-hook-form';
-import { SportType } from '@/shared/model/sportTypes';
-import { OfficialStageData } from '@/shared/types/stage/create/official';
-
-interface ValidateCurrentSportFieldsProps {
-  selectedSport: SportType | null;
-  fields: Array<{
-    id: string;
-    category: SportType;
-  }>;
-  watch: UseFormWatch<OfficialStageData>;
-}
-
-export const validateCurrentSportFields = ({
-  selectedSport,
-  fields,
-  watch,
-}: ValidateCurrentSportFieldsProps): boolean => {
-  if (!selectedSport) return true;
-
-  const currentFields = fields.filter(
-    (field) => field.category === selectedSport,
-  );
-  if (currentFields.length === 0) return true;
-
-  for (let i = 0; i < currentFields.length; i++) {
-    const fieldIndex = fields.findIndex(
-      (field) => field.id === currentFields[i].id,
-    );
-    const fieldName = watch(`game.${fieldIndex}.name`);
-    const fieldMinCapacity = watch(`game.${fieldIndex}.teamMinCapacity`);
-    const fieldMaxCapacity = watch(`game.${fieldIndex}.teamMaxCapacity`);
-
-    if (!fieldName || !fieldMinCapacity || !fieldMaxCapacity) {
-      return false;
-    }
-  }
-
-  return true;
-};
